chore: move logger to core
