+++ packages/compat/webpack/src/progress/ProgressPlugin.ts
@@ -1,4 +1,5 @@
-import { logger, prettyTime } from '@rsbuild/shared';
+import { logger } from '@rsbuild/core';
+import { prettyTime } from '@rsbuild/shared';
 import webpack from 'webpack';
 import { bus, createFriendlyPercentage } from './helpers';
 import { createNonTTYLogger } from './helpers/nonTty';
