+++ packages/core/bin/rsbuild.js
@@ -1,6 +1,5 @@
 #!/usr/bin/env node
-import { logger } from '@rsbuild/shared';
-import { __internalHelper } from '../dist/index.js';
+import { __internalHelper, logger } from '../dist/index.js';
 
 const { runCli, prepareCli } = __internalHelper;
