+++ packages/core/package.json
@@ -72,6 +72,7 @@
     "dotenv-expand": "11.0.6",
     "fs-extra": "^11.2.0",
     "http-compression": "1.0.20",
+    "jiti": "^1.21.6",
     "launch-editor-middleware": "^2.7.0",
     "on-finished": "2.4.1",
     "open": "^8.4.0",
@@ -79,10 +80,10 @@
     "postcss-loader": "8.1.1",
     "postcss-value-parser": "4.2.0",
     "prebundle": "1.1.0",
+    "rslog": "^1.2.2",
     "rspack-manifest-plugin": "5.0.0",
-    "sirv": "^2.0.4",
-    "jiti": "^1.21.6",
     "semver": "^7.6.2",
+    "sirv": "^2.0.4",
     "style-loader": "3.3.4",
     "tsc-alias": "^1.8.10",
     "typescript": "^5.4.2",
