+++ packages/core/src/cli/commands.ts
@@ -1,8 +1,9 @@
 import { existsSync } from 'node:fs';
 import { join } from 'node:path';
-import { type RsbuildMode, color, logger } from '@rsbuild/shared';
+import { type RsbuildMode, color } from '@rsbuild/shared';
 import { type Command, program } from 'commander';
 import { isEmptyDir } from '../helpers';
+import { logger } from '../logger';
 import { init } from './init';
 
 export type CommonOptions = {
