+++ packages/core/src/plugins/open.ts
@@ -1,8 +1,9 @@
 import { exec } from 'node:child_process';
 import { promisify } from 'node:util';
-import { type Routes, castArray, logger } from '@rsbuild/shared';
+import { type Routes, castArray } from '@rsbuild/shared';
 import { STATIC_PATH } from '../constants';
 import { canParse } from '../helpers';
+import { logger } from '../logger';
 import type { NormalizedConfig, RsbuildPlugin } from '../types';
 
 const execAsync = promisify(exec);
