+++ packages/core/src/server/helper.ts
@@ -2,7 +2,7 @@ import type { IncomingMessage } from 'node:http';
 import net from 'node:net';
 import type { Socket } from 'node:net';
 import os from 'node:os';
-import { color, deepmerge, isFunction, logger } from '@rsbuild/shared';
+import { color, deepmerge, isFunction } from '@rsbuild/shared';
 import type {
   DevConfig,
   NormalizedConfig,
@@ -12,6 +12,7 @@ import type {
   RsbuildEntry,
 } from '@rsbuild/shared';
 import { DEFAULT_DEV_HOST, DEFAULT_PORT } from '../constants';
+import { logger } from '../logger';
 
 /**
  * It used to subscribe http upgrade event
