+++ packages/plugin-eslint/package.json
@@ -37,8 +37,8 @@
     "typescript": "^5.4.2"
   },
   "peerDependencies": {
-    "eslint": "^8.0.0 || ^9.0.0",
-    "@rsbuild/core": "workspace:^0.7.8"
+    "@rsbuild/core": "workspace:^0.7.8",
+    "eslint": "^8.0.0 || ^9.0.0"
   },
   "publishConfig": {
     "access": "public",
