+++ packages/plugin-vue2/src/index.ts
@@ -88,9 +88,7 @@ export function pluginVue2(options: PluginVueOptions = {}): RsbuildPlugin {
 
         chain.plugin(CHAIN_ID.PLUGIN.VUE_LOADER_PLUGIN).use(VueLoaderPlugin);
         // we could remove this once a new vue-loader@15 is released with https://github.com/vuejs/vue-loader/pull/2071 shipped
-        chain
-          .plugin(CHAIN_ID.PLUGIN.VUE_LOADER_15_PITCH_FIX_PLUGIN)
-          .use(VueLoader15PitchFixPlugin);
+        chain.plugin('vue-loader-15-pitch-fix').use(VueLoader15PitchFixPlugin);
       });
 
       applySplitChunksRule(api, options.splitChunks);
