+++ packages/shared/src/index.ts
@@ -5,7 +5,6 @@ export { fse };
 export * from './constants';
 export * from './utils';
 export * from './getBrowserslist';
-export * from './logger';
 export * from './reduceConfigs';
 export * from './types';
 export * from './chain';
